third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/webusb-origin-trial-interfaces-script-added-expected.txt

CONSOLE MESSAGE: line 31: Interfaces before adding trial token
interface Navigator
    getter usb
interface USBAlternateInterface
    getter alternateSetting
    getter endpoints
    getter interfaceClass
    getter interfaceName
    getter interfaceProtocol
    getter interfaceSubclass
    method constructor
interface USBConfiguration
    getter configurationName
    getter configurationValue
    getter interfaces
    method constructor
interface USBConnectionEvent
    getter device
    method constructor
interface USBDevice
    getter configuration
    getter configurations
    getter deviceClass
    getter deviceProtocol
    getter deviceSubclass
    getter deviceVersionMajor
    getter deviceVersionMinor
    getter deviceVersionSubminor
    getter manufacturerName
    getter opened
    getter productId
    getter productName
    getter serialNumber
    getter usbVersionMajor
    getter usbVersionMinor
    getter usbVersionSubminor
    getter vendorId
    method claimInterface
    method clearHalt
    method close
    method constructor
    method controlTransferIn
    method controlTransferOut
    method isochronousTransferIn
    method isochronousTransferOut
    method open
    method releaseInterface
    method reset
    method selectAlternateInterface
    method selectConfiguration
    method transferIn
    method transferOut
interface USBEndpoint
    getter direction
    getter endpointNumber
    getter packetSize
    getter type
    method constructor
interface USBInTransferResult
    getter data
    getter status
    method constructor
interface USBInterface
    getter alternate
    getter alternates
    getter claimed
    getter interfaceNumber
    method constructor
interface USBIsochronousInTransferPacket
    getter data
    getter status
    method constructor
interface USBIsochronousInTransferResult
    getter data
    getter packets
    method constructor
interface USBIsochronousOutTransferPacket
    getter bytesWritten
    getter status
    method constructor
interface USBIsochronousOutTransferResult
    getter packets
    method constructor
interface USBOutTransferResult
    getter bytesWritten
    getter status
    method constructor
CONSOLE MESSAGE: line 53: Interfaces after adding trial token
interface Navigator
    getter usb
interface USBAlternateInterface
    getter alternateSetting
    getter endpoints
    getter interfaceClass
    getter interfaceName
    getter interfaceProtocol
    getter interfaceSubclass
    method constructor
interface USBConfiguration
    getter configurationName
    getter configurationValue
    getter interfaces
    method constructor
interface USBConnectionEvent
    getter device
    method constructor
interface USBDevice
    getter configuration
    getter configurations
    getter deviceClass
    getter deviceProtocol
    getter deviceSubclass
    getter deviceVersionMajor
    getter deviceVersionMinor
    getter deviceVersionSubminor
    getter manufacturerName
    getter opened
    getter productId
    getter productName
    getter serialNumber
    getter usbVersionMajor
    getter usbVersionMinor
    getter usbVersionSubminor
    getter vendorId
    method claimInterface
    method clearHalt
    method close
    method constructor
    method controlTransferIn
    method controlTransferOut
    method isochronousTransferIn
    method isochronousTransferOut
    method open
    method releaseInterface
    method reset
    method selectAlternateInterface
    method selectConfiguration
    method transferIn
    method transferOut
interface USBEndpoint
    getter direction
    getter endpointNumber
    getter packetSize
    getter type
    method constructor
interface USBInTransferResult
    getter data
    getter status
    method constructor
interface USBInterface
    getter alternate
    getter alternates
    getter claimed
    getter interfaceNumber
    method constructor
interface USBIsochronousInTransferPacket
    getter data
    getter status
    method constructor
interface USBIsochronousInTransferResult
    getter data
    getter packets
    method constructor
interface USBIsochronousOutTransferPacket
    getter bytesWritten
    getter status
    method constructor
interface USBIsochronousOutTransferResult
    getter packets
    method constructor
interface USBOutTransferResult
    getter bytesWritten
    getter status
    method constructor
This is a testharness.js-based test.
PASS WebUSB related interfaces before adding trial token via script. 
PASS WebUSB related interfaces after adding trial token via script. 
Harness: the test ran to completion.
